requirements.txt
streamlit==1.32.0
torch==2.2.0+cu118
torchvision==0.17.0+cu118
diffusers==0.27.2
transformers==4.37.2
Pillow==10.2.0
opencv-python==4.9.0.80
numpy==1.24.4
controlnet_aux==0.0.7
accelerate==0.27.2 